Created 2 More pages
Firstly, I shifted and created 2 new pages that were more in-depth and help the website more. I've been struggling with the footers and they block some of the text at the bottom so I commented them out for now. I added a button style to the CCS page and added and linked the 2 pages to the main page. I think this is a good start but definitely can be improved on. If there's anything I should add or change please let me know.